# src/GroundControl.Api/OpenApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GroundControl.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/roles/{id}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "summary": "Get a role",
        "description": "Returns a role by its unique identifier. Includes an ETag header for optimistic concurrency.",
        "operationId": "GetRoleHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Roles"
        ],
        "summary": "Update a role",
        "description": "Updates an existing role. Requires an If-Match header with the current ETag value.",
        "operationId": "UpdateRoleHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Roles"
        ],
        "summary": "Delete a role",
        "description": "Deletes a role if it is not assigned to any users. Requires an If-Match header.",
        "operationId": "DeleteRoleHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/scopes/{id}": {
      "get": {
        "tags": [
          "Scopes"
        ],
        "summary": "Get a scope",
        "description": "Returns a scope by its unique identifier. Includes an ETag header for optimistic concurrency.",
        "operationId": "GetScopeHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Scopes"
        ],
        "summary": "Update a scope",
        "description": "Updates an existing scope. Requires an If-Match header with the current ETag value.",
        "operationId": "UpdateScopeHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScopeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Scopes"
        ],
        "summary": "Delete a scope",
        "description": "Deletes a scope if it is not referenced by any configuration entries. Requires an If-Match header.",
        "operationId": "DeleteScopeHandler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "Precondition Required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/snapshots/{id}": {
      "get": {
        "tags": [
          "Snapshots"
        ],
        "summary": "Get a snapshot",
        "description": "Returns a snapshot by its unique identifier. Optionally decrypts sensitive values.",
        "operationId": "GetSnapshotHandler",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "decrypt",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ResolvedEntryResponse": {
        "required": [
          "key",
          "valueType",
          "isSensitive",
          "values"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Gets the configuration key."
          },
          "valueType": {
            "type": "string",
            "description": "Gets the serialized value type name."
          },
          "isSensitive": {
            "type": "boolean",
            "description": "Gets a value indicating whether the entry contains sensitive data."
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScopedValueResponse"
            },
            "description": "Gets the resolved values for each applicable scope combination."
          }
        },
        "description": "Represents a resolved configuration entry in a snapshot response."
      },
      "RoleResponse": {
        "required": [
          "id",
          "name",
          "permissions",
          "version",
          "createdAt",
          "createdBy",
          "updatedAt",
          "updatedBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets the unique identifier for the role.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Gets the display name for the role."
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the optional human-readable description for the role."
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets the permission strings granted by the role."
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Gets the optimistic concurrency version.",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "description": "Gets the timestamp when the role was created.",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "description": "Gets the identifier of the user that created the role.",
            "format": "uuid"
          },
          "updatedAt": {
            "type": "string",
            "description": "Gets the timestamp when the role was last updated.",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "description": "Gets the identifier of the user that last updated the role.",
            "format": "uuid"
          }
        },
        "description": "Represents the API response body for a role."
      },
...
      "ScopedValueResponse": {
        "required": [
          "scopes",
          "value"
        ],
        "type": "object",
        "properties": {
          "scopes": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Gets the scope dimension-value pairs that qualify this value."
          },
          "value": {
            "type": "string",
            "description": "Gets the serialized value for the scope combination."
          }
        },
        "description": "Represents a scope-specific value in a snapshot response."
      },
      "ScopeResponse": {
        "required": [
          "id",
          "dimension",
          "allowedValues",
          "version",
          "createdAt",
          "createdBy",
          "updatedAt",
          "updatedBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets the unique identifier for the scope.",
            "format": "uuid"
          },
          "dimension": {
            "type": "string",
            "description": "Gets the unique scope dimension name."
          },
          "allowedValues": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets the allowed values for the scope dimension."
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the optional human-readable description for the scope."
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Gets the optimistic concurrency version.",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "description": "Gets the timestamp when the scope was created.",
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "description": "Gets the identifier of the user that created the scope.",
            "format": "uuid"
          },
          "updatedAt": {
            "type": "string",
            "description": "Gets the timestamp when the scope was last updated.",
            "format": "date-time"
          },
          "updatedBy": {
            "type": "string",
            "description": "Gets the identifier of the user that last updated the scope.",
            "format": "uuid"
          }
        },
        "description": "Represents the API response body for a scope definition."
      },
...
      "SnapshotResponse": {
        "required": [
          "id",
          "projectId",
          "snapshotVersion",
          "entries",
          "publishedAt",
          "publishedBy"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets the unique identifier for the snapshot.",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "description": "Gets the project identifier that owns the snapshot.",
            "format": "uuid"
          },
          "snapshotVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Gets the monotonically increasing snapshot version within the project.",
            "format": "int64"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResolvedEntryResponse"
            },
            "description": "Gets the resolved entries captured in the snapshot."
          },
          "publishedAt": {
            "type": "string",
            "description": "Gets the publication timestamp.",
            "format": "date-time"
          },
          "publishedBy": {
            "type": "string",
            "description": "Gets the identifier of the publishing user.",
            "format": "uuid"
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the optional publication description."
          }
        },
        "description": "Represents the full API response body for a snapshot, including resolved entries."
      },
...
      "UpdateRoleRequest": {
        "required": [
          "name",
          "permissions"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "Gets the display name for the role."
          },
          "description": {
            "maxLength": 500,
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the optional human-readable description for the role."
          },
          "permissions": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets the permission strings granted by the role."
          }
        },
        "description": "Represents the request body for updating a role."
      },
      "UpdateScopeRequest": {
        "required": [
          "dimension",
          "allowedValues"
        ],
        "type": "object",
        "properties": {
          "dimension": {
            "maxLength": 100,
            "type": "string",
            "description": "Gets the unique scope dimension name."
          },
          "allowedValues": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Gets the allowed values for the scope dimension."
          },
          "description": {
            "maxLength": 500,
            "type": [
              "null",
              "string"
            ],
            "description": "Gets the optional human-readable description for the scope."
          }
        },
        "description": "Represents the request body for updating a scope definition."
      },
...
    }
  },
...
}